+++ axelor-base/src/main/java/com/axelor/apps/base/exceptions/IExceptionMessage.java
@@ -387,4 +387,15 @@ public interface IExceptionMessage {
   public static final String ADVANCED_IMPORT_NO_RESET = /*$$(*/ "No imported data to reset" /*)*/;
   public static final String SERVER_CONNECTION_ERROR = /*$$(*/
       "Unable to connect to Geonames server" /*)*/;
+
+  /** External Report */
+  public static final String TEMPLATE_FIELD_EMPTY = /*$$(*/ "Please select birt template." /*)*/;
+
+  public static final String MISSING_TEMPLATE_FILE = /*$$(*/
+      "Missing template file on birt template." /*)*/;
+  public static final String META_MODEL_NOT_FOUND = /*$$(*/
+      "There is no birt template to print report for %s" /*)*/;
+  public static final String MISSING_BIRT_TEMPLATE_ID = /*$$(*/
+      "Id field is missing please select birt template." /*)*/;
+  public static final String MISSING_RECORD_ID = /*$$(*/ "Please save this record." /*)*/;
 }